==> code/01_prep_candidates.Rmd <==
---
title: "Data preprocessing - candidate dataset"
author: "Silvia Téliz"
date: "2025-03-25"
output: html_document
---

This script carries out the preprocessing of the data for candidates in the 2024 elections. Only proprietary and representative majority candidates are included.

# Setup

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```


```{r packages}
library(tidyverse)
library(magrittr)
library(readxl)
library(stringi)
```

# Import data
```{r data}
# Import original candidate dataset
all_candidates <- read_excel("../data/raw/baseDatosCandidatos.xls")

```

# General wrangling
```{r wrang_general}
# Exclude irrelevant columns
all_candidates %<>% select(-c("CIRCUNSCRIPCION", "MUNICIPIO", "NUM_LISTA_O_FORMULA", "DIRECCION_CASA_CAMPAÑA", "TELEFONO", "CORREO_ELECTRONICO", "ESCOLARIDAD", "ESTATUS_ESCOLARIDAD", "CURSOS", "HISTORIA_PROFESIONAL", "TRAYECTORIA_POLITICA", "MOTIVO_CARGO_PUBLICO"))


names(all_candidates) <- c("party", "position", "state", "district", "name", "sv_pp", "ine_profile", "type", "estatus", "age", "sex", "webpage", "social_media", "prop_1", "prop_2", "prop_gender")

# Turn all content to lowercase except urls
all_candidates %<>% 
  mutate(across(c(party, position, state, district, name, type, estatus, age,
                  sex, webpage, social_media, prop_1, prop_2, prop_gender), tolower))

# Remove punctuation from name column
all_candidates$name <- str_remove_all(all_candidates$name, "[[:punct:]]")

# Recode sex
all_candidates$sex %<>% fct_recode("f" = "mujer", "m" = "hombre", "nb" = "no binario")

# Recode position
all_candidates$position %<>% fct_recode("pres" = "presidencia de la república",
                                        "dep" = "diputación federal mr",
                                        "sen" = "senaduría federal mr")
```

```{r ine_pp}
# Extract the number in "ine_profile" that lies in between "detallecandidato/" and "/"
all_candidates$candidate_num <- as.numeric(str_extract(all_candidates$ine_profile, "(?<=detalleCandidato\\/)\\d*(?=\\/\\d)"))

# Create a new column "ine_pp" 
all_candidates$ine_pp <- paste0("https://candidaturas.ine.mx/cycc/img/fotocandidato/", all_candidates$candidate_num, ".jpg")

# Remove candidate_num column
all_candidates <- all_candidates %>% select(-candidate_num)

```


```{r other_dummy}
# Create a portrait column that indicates whether a candidate has a portrait or not
all_candidates$sv_d <- ifelse(is.na(all_candidates$sv_pp), 0, 1)
```


```{r ine_dummy}
# Create a portrait column that indicates whether a candidate has a portrait or not
all_candidates$ine_d <- ifelse(all_candidates$estatus=="sin captura", 0, 1)

# Remove estatus column
all_candidates <- all_candidates %>% select(-estatus)
```


```{r webpage_dummy}
# Create webpage column that indicates whether a candidate has a webpage or not. If pagina_web is NA or "no proporcionó", then webpage is 0, otherwise 1.
all_candidates$webpage_d <- ifelse(is.na(all_candidates$webpage) | all_candidates$webpage == "no proporcionó", 0, 1)

# Replace "no proporcionó" with NA
all_candidates$webpage <- ifelse(all_candidates$webpage == "no proporcionó", NA, all_candidates$webpage)

```


```{r socialmedia_dummy}
# Create socialmedia column that indicates whether a candidate has social media or not. If redes is NA or "no proporcionó", then socialmedia is 0, otherwise 1.
all_candidates$social_media_d <- ifelse(is.na(all_candidates$social_media) | all_candidates$social_media == "no proporcionó", 0, 1)

# Replace "no proporcionó" with NA
all_candidates$social_media <- ifelse(all_candidates$social_media == "no proporcionó", NA, all_candidates$social_media)

```


```{r sm_d}
# Create separate columns for facebook, instagram, twitter, youtube, and linkedin urls from content in sm_1, sm_2, sm_3, sm_4, and sm_5 columns.

all_candidates$facebook <- str_extract(all_candidates$social_media, "https://www.facebook.com/[^,]+")

all_candidates$instagram <- str_extract(all_candidates$social_media, "https://www.instagram.com/[^,]+")

all_candidates$twitter <- str_extract(all_candidates$social_media, "https://twitter.com/[^,]+")

all_candidates$youtube <- str_extract(all_candidates$social_media, "https://www.youtube.com/[^,]+")

all_candidates$tiktok <- str_extract(all_candidates$social_media, "https://www.tiktok.com/[^,]+")
```



```{r state_id}
# Create a state_id column
all_candidates$state_id <- all_candidates$state 
  
all_candidates$state_id %<>%
           fct_recode("1" = "aguascalientes",
                      "2" = "baja california",
                      "3" = "baja california sur",
                      "4" = "campeche",
                      "5" = "coahuila",
                      "6" = "colima",
                      "7" = "chiapas",
                      "8" = "chihuahua",
                      "9" = "ciudad de mexico",
                      "10" = "durango",
                      "11" = "guanajuato",
                      "12" = "guerrero",
                      "13" = "hidalgo",
                      "14" = "jalisco",
                      "15" = "mexico",
                      "16" = "michoacan",
                      "17" = "morelos",
                      "18" = "nayarit",
                      "19" = "nuevo leon",
                      "20" = "oaxaca",
                      "21" = "puebla",
                      "22" = "queretaro",
                      "23" = "quintana roo",
                      "24" = "san luis potosi",
                      "25" = "sinaloa",
                      "26" = "sonora",
                      "27" = "tabasco",
                      "28" = "tamaulipas",
                      "29" = "tlaxcala",
                      "30" = "veracruz",
                      "31" = "yucatan",
                      "32" = "zacatecas",
                      "33" = "nacional")



#summary(as_factor(all_candidates$state_id))
```

```{r district_id}
# Create a district_id column that extracts the number from the district column. In case the content of district is "no aplica", the value of district_id will be 0.
all_candidates <- all_candidates %>%
  mutate(district_id = ifelse(str_detect(district, "no aplica"), 
                              0, 
                              as.numeric(str_extract(district, "\\d+"))))
```

```{r state_district}
# Create a state_district column that concatenates the state_id and district_id columns
all_candidates$state_district <- paste(all_candidates$state_id, all_candidates$district_id, sep="_")
```


```{r party_single}
# Create dummy variables to determine candidates' party affiliation
# summary(as_factor(all_candidates$party))

# Create MORENA column
all_candidates$morena <- ifelse(str_detect(all_candidates$party, "morena|sigamos haciendo historia"), 1, 0)

# Create PAN column
all_candidates$pan <- ifelse(str_detect(all_candidates$party, "partido acción nacional|fuerza y corazon por mexico"), 1, 0)

# Create PRI column
all_candidates$pri <- ifelse(str_detect(all_candidates$party, "partido revolucionario institucional|fuerza y corazon por mexico"), 1, 0)

# Create PT column
all_candidates$pt <- ifelse(str_detect(all_candidates$party, "partido del trabajo|sigamos haciendo historia"), 1, 0)

# Create PVEM column
all_candidates$pvem <- ifelse(str_detect(all_candidates$party, "partido verde ecologista de méxico|sigamos haciendo historia"), 1, 0)

# Create MC column
all_candidates$mc <- ifelse(str_detect(all_candidates$party, "movimiento ciudadano"), 1, 0)

# Create PRD column
all_candidates$prd <- ifelse(str_detect(all_candidates$party, "partido de la revolución democrática|fuerza y corazon por mexico"), 1, 0)

# Create IND column
all_candidates$ind <- ifelse(str_detect(all_candidates$party, "candidatura independiente"), 1, 0)

```

```{r recode_party}
# Recode party column
all_candidates$party %<>%
  fct_recode("morena" = "morena",
             "pan" = "partido acción nacional",
             "pri" = "partido revolucionario institucional",
             "pt" = "partido del trabajo",
             "pvem" = "partido verde ecologista de méxico",
             "mc" = "movimiento ciudadano",
             "prd" = "partido de la revolución democrática",
             "ind" = "candidatura independiente",
             "pvem_pt_morena" = "sigamos haciendo historia",
             "pan_pri_prd" = "fuerza y corazon por mexico")
```


```{r id}
# Create a id column that concatenates the state_id, district_id, and the first five letters of the name column.
all_candidates <- all_candidates %>%
  mutate(id = paste0(state_id, district_id, substr(name, 1, 4), substr(party, 1, 3)))

# Remove blank spaces in id
all_candidates$id <- str_replace_all(all_candidates$id, " ", "")

```

# Create dataset per position

## Chamber of deputies

```{r dep_candidates}
dep_candidates <- all_candidates %>% filter(position == "dep" & type == "persona propietaria")

dep_candidates %>% count(id) %>% filter(n > 1)
```


## Senate
```{r sen_candidates}
sen_candidates <- all_candidates %>% filter(position == "sen" & type == "persona propietaria")

sen_candidates %>% count(id) %>% filter(n > 1)
```


## Presidency

```{r sen_candidates}
pres_candidates <- all_candidates %>% filter(position == "pres" & type == "persona propietaria")

pres_candidates %>% count(id) %>% filter(n > 1)
```


# Add incumbency variable
```{r wrang_incumbency}
# Import list of previous congress and senate (LXV legislature)
dep_65 <- read.csv("../data/raw/deputies_65.csv")
sen_65 <- read.csv("../data/raw/senators_65.csv")

# Convert name column to lowercase in dep_65 and sen_65
dep_65$name <- tolower(dep_65$name)
sen_65$name <- tolower(sen_65$name)

# Remove punctuation from name column
dep_65$name <- str_remove_all(dep_65$name, "[[:punct:]]")
sen_65$name <- str_remove_all(sen_65$name, "[[:punct:]]")

# Remove accents from name column
dep_65$name <- stri_trans_general(dep_65$name, "Latin-ASCII")
sen_65$name <- stri_trans_general(sen_65$name, "Latin-ASCII")

# check uniqueness of name columns
dep_candidates %>% count(name) %>% filter(n > 1)
sen_candidates %>% count(name) %>% filter(n > 1)
pres_candidates %>% count(name) %>% filter(n > 1)

dep_65 %>% count(name) %>% filter(n > 1)
sen_65 %>% count(name) %>% filter(n > 1)
```


```{r incumbency}
# Create incumbency variable in dep_candidates. If name in dep_candidates is present in name in dep_65, then incumbency is 1, otherwise 0.
dep_candidates %<>% 
  mutate(incumbency = ifelse(name %in% dep_65$name, 1, 0))

# Create incumbency for sen_candidates
sen_candidates %<>% 
  mutate(incumbency = ifelse(name %in% sen_65$name, 1, 0))

# Count incumbency
dep_candidates %>% count(incumbency)
sen_candidates %>% count(incumbency)

# Create incumbency for pres_candidates
pres_candidates %<>% 
  mutate(incumbency = 0)
```

# Join datasets

```{r join}
# Join datasets
candidates <- bind_rows(dep_candidates, sen_candidates, pres_candidates)

# Remove "type" column
candidates <- candidates %>% select(-type)

# Place id and name columns at the beginning
candidates <- candidates %>% select(id, name, party, position, state, state_id, district, district_id, state_district, incumbency, everything())
```


# Export dataset

```{r export}
# Export dataset
write_csv(candidates, "../data/processed/01_candidates.csv")
```
