==> code/03_prep_elections.Rmd <==
---
title: "Data preprocessing - elections dataset"
author: "Silvia Téliz"
date: "2025-05-12"
output: html_document
---

This script contains the data preprocessing steps for the elections dataset. 

# Setup

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r packages}
library(tidyverse)
library(magrittr)
library(purrr)
```

# Import data

```{r data}

# Candidates dataset
candidates <- read_csv("../data/processed/01_candidates.csv") %>%
  select(-c(morena, pan, pri, pt, pvem, mc, prd, ind))

# Election results
dep <- read_csv("../data/raw/election_results/DIP_FED_2024.csv") %>% 
  select(-c("CLAVE_CASILLA", "SECCION", "ID_CASILLA", "TIPO_CASILLA", "EXT_CONTIGUA", "CASILLA", "TIPO_ACTA", "OBSERVACIONES", "MECANISMOS_TRASLADO", "FECHA_HORA"))


sen <- read_csv("../data/raw/election_results/SEN_2024.csv")%>% 
  select(-c("CLAVE_CASILLA", "SECCION", "ID_CASILLA", "TIPO_CASILLA", "EXT_CONTIGUA", "CASILLA", "TIPO_ACTA", "OBSERVACIONES", "MECANISMOS_TRASLADO", "FECHA_HORA"))


pres <- read_csv("../data/raw/election_results/PRES_2024.csv")%>% 
  select(-c("CLAVE_CASILLA", "SECCION", "ID_CASILLA", "TIPO_CASILLA", "EXT_CONTIGUA", "CASILLA", "TIPO_ACTA", "OBSERVACIONES", "MECANISMOS_TRASLADO", "FECHA_HORA"))
```

# General wrangling

```{r clean_ids}
# Remove zeros at the begining of strings in ID_ENTIDAD and ID_DISTRITO_FEDERAL
dep <- dep %>% 
  mutate(ID_ENTIDAD = str_remove(ID_ENTIDAD, "^0+"),
         ID_DISTRITO_FEDERAL = str_remove(ID_DISTRITO_FEDERAL, "^0+"))

sen <- sen %>%
  mutate(ID_ENTIDAD = str_remove(ID_ENTIDAD, "^0+"),
         ID_DISTRITO_FEDERAL = str_remove(ID_DISTRITO_FEDERAL, "^0+"))

pres <- pres %>%
  mutate(ID_ENTIDAD = str_remove(ID_ENTIDAD, "^0+"),
         ID_DISTRITO_FEDERAL = str_remove(ID_DISTRITO_FEDERAL, "^0+"))

# Tranform N/A string to 0 in ID_ENTIDAD and ID_DISTRITO_FEDERAL
dep <- dep %>% 
  mutate(ID_ENTIDAD = ifelse(ID_ENTIDAD == "N/A", "0", ID_ENTIDAD),
         ID_DISTRITO_FEDERAL = ifelse(ID_DISTRITO_FEDERAL == "N/A", "0", ID_DISTRITO_FEDERAL))

sen <- sen %>%
  mutate(ID_ENTIDAD = ifelse(ID_ENTIDAD == "N/A", "0", ID_ENTIDAD),
         ID_DISTRITO_FEDERAL = ifelse(ID_DISTRITO_FEDERAL == "N/A", "0", ID_DISTRITO_FEDERAL))


pres <- pres %>%
  mutate(ID_ENTIDAD = ifelse(ID_ENTIDAD == "N/A", "0", ID_ENTIDAD),
         ID_DISTRITO_FEDERAL = ifelse(ID_DISTRITO_FEDERAL == "N/A", "0", ID_DISTRITO_FEDERAL))
```

```{r state_district}
# Create state_district column
dep <- dep %>% 
  mutate(state_district = paste(ID_ENTIDAD, ID_DISTRITO_FEDERAL, sep="_"))

sen <- sen %>%
  mutate(state_district = paste(ID_ENTIDAD, ID_DISTRITO_FEDERAL, sep="_"))

pres <- pres %>%
  mutate(state_district = paste(ID_ENTIDAD, ID_DISTRITO_FEDERAL, sep="_"))


# Place state_district column after DISTRITO_FEDERAL
dep <- dep %>% 
  relocate(state_district, .after = DISTRITO_FEDERAL)

sen <- sen %>%
  relocate(state_district, .after = DISTRITO_FEDERAL)

pres <- pres %>%
  relocate(state_district, .after = DISTRITO_FEDERAL)

# Transform to numeric all columns from PAN to LISTA_NOMINAL
dep <- dep %>% 
  mutate(across(PAN:LISTA_NOMINAL, ~ as.numeric(.x)))

sen <- sen %>%
  mutate(across(PAN:LISTA_NOMINAL, ~ as.numeric(.x)))

pres <- pres %>%
  mutate(across(PAN:LISTA_NOMINAL, ~ as.numeric(.x)))

```

# Generate dataset per contest

```{r deputies}
dep_results <- dep %>%
  group_by(state_district) %>%
  summarize(
    pan = sum(PAN, na.rm = TRUE),
    pri = sum(PRI, na.rm = TRUE),
    prd = sum(PRD, na.rm = TRUE),
    pvem = sum(PVEM, na.rm = TRUE),
    pt = sum(PT, na.rm = TRUE),
    mc = sum(MC, na.rm = TRUE),
    morena = sum(MORENA, na.rm = TRUE),
    ind = sum(IND, na.rm = TRUE),
    pan_pri_prd = sum(PAN_PRI_PRD, na.rm = TRUE),
    pan_pri = sum(PAN_PRI, na.rm = TRUE),
    pan_prd = sum(PAN_PRD, na.rm = TRUE),
    pri_prd = sum(PRI_PRD, na.rm = TRUE),
    pvem_pt_morena = sum(PVEM_PT_MORENA, na.rm = TRUE),
    pvem_pt = sum(PVEM_PT, na.rm = TRUE),
    pvem_morena = sum(PVEM_MORENA, na.rm = TRUE),
    pt_morena = sum(PT_MORENA, na.rm = TRUE),
    total_counted = sum(TOTAL_VOTOS_CALCULADOS, na.rm = TRUE),
    total_registered = sum(LISTA_NOMINAL, na.rm = TRUE))

```


```{r senators}
sen_results <- sen %>%
  group_by(ID_ENTIDAD) %>%
  summarize(
    pan = sum(PAN, na.rm = TRUE),
    pri = sum(PRI, na.rm = TRUE),
    prd = sum(PRD, na.rm = TRUE),
    pvem = sum(PVEM, na.rm = TRUE),
    pt = sum(PT, na.rm = TRUE),
    mc = sum(MC, na.rm = TRUE),
    morena = sum(MORENA, na.rm = TRUE),
    pan_pri_prd = sum(PAN_PRI_PRD, na.rm = TRUE),
    pan_pri = sum(PAN_PRI, na.rm = TRUE),
    pan_prd = sum(PAN_PRD, na.rm = TRUE),
    pri_prd = sum(PRI_PRD, na.rm = TRUE),
    pvem_pt_morena = sum(PVEM_PT_MORENA, na.rm = TRUE),
    pvem_pt = sum(PVEM_PT, na.rm = TRUE),
    pvem_morena = sum(PVEM_MORENA, na.rm = TRUE),
    pt_morena = sum(PT_MORENA, na.rm = TRUE),
    total_counted = sum(TOTAL_VOTOS_CALCULADOS, na.rm = TRUE),
    total_registered = sum(LISTA_NOMINAL, na.rm = TRUE))

# Rename ID_ENTIDAD to state
sen_results <- sen_results %>%
  rename(state_id = ID_ENTIDAD) 
sen_results$state_id %<>% as.numeric()
```



```{r president}
pres_results <- pres %>%
  group_by(ID_ENTIDAD) %>%
  summarize(
    pan = sum(PAN, na.rm = TRUE),
    pri = sum(PRI, na.rm = TRUE),
    prd = sum(PRD, na.rm = TRUE),
    pvem = sum(PVEM, na.rm = TRUE),
    pt = sum(PT, na.rm = TRUE),
    mc = sum(MC, na.rm = TRUE),
    morena = sum(MORENA, na.rm = TRUE),
    pan_pri_prd = sum(PAN_PRI_PRD, na.rm = TRUE),
    pan_pri = sum(PAN_PRI, na.rm = TRUE),
    pan_prd = sum(PAN_PRD, na.rm = TRUE),
    pri_prd = sum(PRI_PRD, na.rm = TRUE),
    pvem_pt_morena = sum(PVEM_PT_MORENA, na.rm = TRUE),
    pvem_pt = sum(PVEM_PT, na.rm = TRUE),
    pvem_morena = sum(PVEM_MORENA, na.rm = TRUE),
    pt_morena = sum(PT_MORENA, na.rm = TRUE),
    total_counted = sum(TOTAL_VOTOS_CALCULADOS, na.rm = TRUE),
    total_registered = sum(LISTA_NOMINAL, na.rm = TRUE)
  )

# Rename ID_ENTIDAD to state
pres_results <- pres_results %>%
  rename(state_id = ID_ENTIDAD)

pres_results$state_id %<>% as.numeric()
```


```{r sep_candidates}
dep_candidates <- candidates %>% 
  filter(position == "dep")

sen_candidates <- candidates %>% 
  filter(position == "sen")

pres_candidates <- candidates %>%
  filter(position == "pres")

```

# Match candidates with results

```{r dep_cand_results}
# Match dep_candidates with dep_results
dep_cand_results <- dep_candidates %>%
  left_join(dep_results, by = "state_district")

#summary(as_factor(dep_cand_results$party))

# Create earned_votes column. If party is in the vote_map list, sum the votes of the parties in the list.
dep_cand_results %<>%
  mutate(earned_votes = case_when(
    party == "pri" ~ rowSums(select(., all_of(c("pri", "pan_pri_prd", "pan_pri", "pri_prd"))), na.rm = TRUE),
    party == "pvem_pt_morena" ~ rowSums(select(., all_of(c("pvem_pt_morena", "pvem", "pt", "morena", "pvem_pt", "pvem_morena", "pt_morena"))), na.rm = TRUE),
    party == "mc" ~ mc,
    party == "pan_pri_prd" ~ rowSums(select(., all_of(c("pan_pri_prd", "pan", "pri", "prd", "pan_pri", "pan_prd", "pri_prd"))), na.rm = TRUE),
    party == "pan" ~ rowSums(select(., all_of(c("pan", "pan_pri_prd", "pan_pri", "pan_prd"))), na.rm = TRUE),
    party == "prd" ~ rowSums(select(., all_of(c("prd", "pan_pri_prd", "pan_prd", "pri_prd"))), na.rm = TRUE),
    party == "morena" ~ rowSums(select(., all_of(c("morena", "pvem_pt_morena", "pvem_morena", "pt_morena"))), na.rm = TRUE),
    party == "pvem" ~ rowSums(select(., all_of(c("pvem", "pvem_pt_morena", "pvem_pt", "pvem_morena"))), na.rm = TRUE),
    party == "pt" ~ rowSums(select(., all_of(c("pt", "pvem_pt_morena", "pvem_pt", "pt_morena"))), na.rm = TRUE),
    party == 'ind' ~ ind,
    TRUE ~ NA_real_
  ))

# Add proportion votes (earned_votes/total_counted)
dep_cand_results %<>%
  mutate(prop_votes = earned_votes / total_counted)

# Add proportion of participation (total_counted/total_registered)
dep_cand_results %<>%
  mutate(prop_participation = total_counted / total_registered)

# Add win dummy variable. The candidate with the highest prop_votes in each state is the winner.
dep_cand_results %<>%
  group_by(state_district) %>%
  mutate(win = ifelse(dense_rank(desc(earned_votes)) <= 1, 1, 0)) %>%
  ungroup()

# Remove party columns
dep_cand_results %<>%
  select(-c(pan, pri, prd, pvem, pt, mc, morena, ind, pan_pri_prd, pan_pri, pan_prd, pri_prd, pvem_pt_morena, pvem_pt, pvem_morena, pt_morena))

```



```{r sen_cand_results}
# Match sen_candidates with sen_results
sen_cand_results <- sen_candidates %>%
  left_join(sen_results, by = "state_id")


#summary(as_factor(sen_cand_results$party))

# Create earned_votes column. If party is in the vote_map list, sum the votes of the parties in the list.
sen_cand_results %<>%
  mutate(earned_votes = case_when(
    party == "pri" ~ rowSums(select(., all_of(c("pri", "pan_pri_prd", "pan_pri", "pri_prd"))), na.rm = TRUE),
    party == "pvem_pt_morena" ~ rowSums(select(., all_of(c("pvem_pt_morena", "pvem", "pt", "morena", "pvem_pt", "pvem_morena", "pt_morena"))), na.rm = TRUE),
    party == "mc" ~ mc,
    party == "pan_pri_prd" ~ rowSums(select(., all_of(c("pan_pri_prd", "pan", "pri", "prd", "pan_pri", "pan_prd", "pri_prd"))), na.rm = TRUE),
    party == "pan" ~ rowSums(select(., all_of(c("pan", "pan_pri_prd", "pan_pri", "pan_prd"))), na.rm = TRUE),
    party == "prd" ~ rowSums(select(., all_of(c("prd", "pan_pri_prd", "pan_prd", "pri_prd"))), na.rm = TRUE),
    party == "morena" ~ rowSums(select(., all_of(c("morena", "pvem_pt_morena", "pvem_morena", "pt_morena"))), na.rm = TRUE),
    party == "pvem" ~ rowSums(select(., all_of(c("pvem", "pvem_pt_morena", "pvem_pt", "pvem_morena"))), na.rm = TRUE),
    party == "pt" ~ rowSums(select(., all_of(c("pt", "pvem_pt_morena", "pvem_pt", "pt_morena"))), na.rm = TRUE),
    TRUE ~ NA_real_
  ))

# Add proportion votes (earned_votes/total_counted)
sen_cand_results %<>%
  mutate(prop_votes = earned_votes / total_counted)

# Add proportion of participation (total_counted/total_registered)
sen_cand_results %<>%
  mutate(prop_participation = total_counted / total_registered)

# Add win dummy variable. The candidate with the highest prop_votes in each state is the winner.
sen_cand_results %<>%
  group_by(state_id) %>%
  mutate(win = ifelse(dense_rank(desc(earned_votes)) <= 1, 1, 0)) %>%
  ungroup()

# Remove party columns
sen_cand_results %<>%
  select(-c(pan, pri, prd, pvem, pt, mc, morena, pan_pri_prd, pan_pri, pan_prd, pri_prd, pvem_pt_morena, pvem_pt, pvem_morena, pt_morena))

```



```{r pres_cand_results}

pres_results <- pres_results %>%
  mutate(sheinbaum = (morena+pvem_pt_morena+pvem_pt+pvem_morena+pvem+pt+pt_morena),
         xochitl = (pan+pri+prd+pan_pri_prd+pan_pri+pan_prd+pri_prd),
         maynez = (mc))

# Precompute vote totals from pres_results
sheinbaum_votes <- sum(pres_results$sheinbaum, na.rm = TRUE)
xochitl_votes <- sum(pres_results$xochitl, na.rm = TRUE)
maynez_votes <- sum(pres_results$maynez, na.rm = TRUE)
total_counted <- sum(pres_results$total_counted, na.rm = TRUE)
total_registered <- sum(pres_results$total_registered, na.rm = TRUE)

# Then assign them by party
pres_cand_results <- pres_candidates %>%
  mutate(earned_votes = case_when(
    party == "pvem_pt_morena" ~ sheinbaum_votes,
    party == "pan_pri_prd" ~ xochitl_votes,
    party == "mc" ~ maynez_votes,
    TRUE ~ NA_real_
  ))

# Add total_counted and total_registered columns
pres_cand_results %<>%
  mutate(total_counted = total_counted,
         total_registered = total_registered)

# Add proportion votes (earned_votes/total_counted)
pres_cand_results %<>%
  mutate(prop_votes = earned_votes / total_counted)

# Add proportion of participation (total_counted/total_registered)
pres_cand_results %<>%
  mutate(prop_participation = total_counted / total_registered)

# Win dummy
pres_cand_results %<>%
  mutate(win = case_when(
    party == "pvem_pt_morena" ~ ifelse(earned_votes == max(earned_votes), 1, 0),
    party == "pan_pri_prd" ~ ifelse(earned_votes == max(earned_votes), 1, 0),
    party == "mc" ~ ifelse(earned_votes == max(earned_votes), 1, 0),
    TRUE ~ NA_real_
  ))

  
```


# Export datasets

```{r export_csv}
# Export the results to CSV files
write_csv(dep_cand_results, "../data/processed/03_dep_cand_results.csv")
write_csv(sen_cand_results, "../data/processed/03_sen_cand_results.csv")
write_csv(pres_cand_results, "../data/processed/03_pres_cand_results.csv")

```
