==> code/04_skintone.ipynb <==
# Skin tone analysis
For this analysis I use the skin-tone-classifier python module (https://github.com/ChenglongMa/SkinToneClassifier?tab=readme-ov-file#install-the-cli-mode-and-the-gui-mode). 

# Import packages
import numpy as np
import pandas as pd
import stone
from json import dumps

## INE portraits

# Obtain skin color measurements from ine portraits
!stone -i ../portraits/ine --debug -o ../data/processed/skintone_ine -t color -p proder -l 1-11

# -o = output directory
# -l = label (1-11)
# -p = PRODER color palette

### Clean INE results data

# Load results data
ine_results = pd.read_csv('../data/processed/skintone_ine/result.csv') 

# Change column names
ine_results = ine_results.rename(columns={'image type': 'image_type', 'face id': 'face_id', 'dominant 1': 'dominant_1',
                            'percent 1':'percent_1', 'dominant 2': 'dominant_2', 'percent 2': 'percent_2',
                            'skin tone': 'skin_tone', 'tone label': 'tone_label'})

# Add column to identify data source
ine_results['source'] = 'ine'

## SaberVotar portraits

# Obtain skin color measurements from sabervotar portraits
!stone -i ../portraits/sv --debug -o ../data/processed/skintone_sv -t color -p proder -l 1-11

# -o = output directory
# -l = label (1-11)
# -p = PRODER color palette

## Clean SaberVotar results data

# Load results data
sv_results = pd.read_csv('../data/processed/skintone_sv/result.csv') 

# Change column names
sv_results = sv_results.rename(columns={'image type': 'image_type', 'face id': 'face_id', 'dominant 1': 'dominant_1',
                            'percent 1':'percent_1', 'dominant 2': 'dominant_2', 'percent 2': 'percent_2',
                            'skin tone': 'skin_tone', 'tone label': 'tone_label'})

# Add column to identify data source
sv_results['source'] = 'sv'

## Export results dataset

# Join ine and sabervotar results
results = pd.concat([ine_results, sv_results], ignore_index=True)

# Export results to CSV
results.to_csv('../data/processed/04_skintone.csv', index=False) 

#
