==> code/05_matching.Rmd <==
---
title: "Descriptive analysis"
author: "Silvia Téliz"
date: "2025-05-13"
output: html_document
---

This script matches the skin tone data with the candidate results data. The skin tone data is obtained from the `skintone` dataset, which contains the skin tone labels for candidates' portraits. The candidate results data is obtained from the `dep_cand_results`, `sen_cand_results`, and `pres_cand_results` datasets. The resulting dataset is `cand_skin_results`, which contains the skin tone data for each candidate, as well as additional variables such as the mean skin tone per state, contest, and party. This is the main dataset used in the analysis.

# Setup

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r packages}
library(tidyverse)
library(magrittr)
library(haven)
```

# Import data

```{r data}
# Candidate-results data
dep_cand_results <- read_csv("../data/processed/03_dep_cand_results.csv")
pres_cand_results <- read_csv("../data/processed/03_pres_cand_results.csv")
sen_cand_results <- read_csv("../data/processed/03_sen_cand_results.csv")

# Skin tone measurement data
skintone <- read_csv("../data/processed/04_skintone.csv")

# PRODER data
proder <- read_dta("../data/raw/baseproder2019.dta")

```

# Skin tone data

```{r clean_skintone}
# Reverse tone_label (1 is 11, 11 is 1). This is necessary because the CASCo algorithm labels skin tones from 1 to 11, where 1 is the darkest skin tone and 11 is the lightest skin tone. 
skintone <- skintone %>%
  mutate(tone_label = 12-tone_label)

# Select relevant variables from skintone data. In some images, two faces were detected. We select the one with the highest confidence score.

skintone_ine <- skintone %>%
  filter(source == "ine" & face_id == 1) %>%
  select(file, tone_label, skin_tone)
names(skintone_ine) <- c("id", "skin_ine", "hex_ine")


skintone_sv <- skintone %>%
  filter(source == "sv" & face_id == 1) %>%
  select(file, tone_label, skin_tone)
names(skintone_sv) <- c("id", "skin_sv", "hex_sv")

# Remove .jpg extension from id column
skintone_ine <- skintone_ine %>%
  mutate(id = str_remove(id, ".jpg"))
skintone_sv <- skintone_sv %>%
  mutate(id = str_remove(id, ".jpg"))

```


```{r match}
# Match skin tone data with candidate results

# Deputies dataset
dep_cand_results <- dep_cand_results %>%
  left_join(skintone_ine, by = "id") %>%
  left_join(skintone_sv, by = "id") %>%
  mutate(contest = "dep")

# Senators dataset
sen_cand_results <- sen_cand_results %>%
  left_join(skintone_ine, by = "id") %>%
  left_join(skintone_sv, by = "id") %>%
  mutate(contest = "sen")

# President dataset
pres_cand_results <- pres_cand_results %>%
  left_join(skintone_ine, by = "id") %>%
  left_join(skintone_sv, by = "id") %>%
  mutate(contest = "pres")

```

# PRODER data
```{r proder}
# PRODER data contains RGB values for skin tones, which we will match with the skin tone data from the candidates. It is necessary to match the RGB values with the skin tone labels from the PRODER palette.

# RGB to HEX dataset
palette <- data.frame(
  hex = c("#654d3e", "#775741", "#876249", "#946c51", "#a0765a", "#a87f64", 
          "#b1886c", "#b69279", "#be9d86", "#c5a691", "#c8ac99"),
  rgb = c("101, 77, 62", "119, 87, 65", "135, 98, 73", "148, 108, 81", "160, 118, 90", "168, 127, 100", 
          "177, 136, 108", "182, 146, 121", "190, 157, 134", "197, 166, 145", "200, 172, 153"),
  label = c(11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1))

# Separate rgb values into different columns in palette dataset
palette %<>%
  separate(rgb, into = c("r", "g", "b"), sep = ", ") %>%
  mutate(r = as.numeric(r), g = as.numeric(g), b = as.numeric(b))

# Select relevant variables
proder %<>%
  select(ent, p160101_codigohex, r_rgbd, g_rgbd, b_rgbd)

# Compute euclidean distance between RGB values and palette
compute_distance <- function(row_rgb, ref_rgb) {
  sqrt(sum((row_rgb - ref_rgb)^2))
}

# For each observation, find the closest category
proder %<>%
  # Remove rows with NA in any column
  filter(!is.na(r_rgbd) | !is.na(g_rgbd) | !is.na(b_rgbd)) %>%
  rowwise() %>%
  mutate(
    proder_label = {
      obs_rgb <- c(r_rgbd, g_rgbd, b_rgbd)
      distances <- apply(palette[, 2:4], 1, function(ref_rgb) compute_distance(obs_rgb, ref_rgb))
      palette$label[which.min(distances)]
    }
  ) %>%
  ungroup()

# Obtain average skin tone per state
proder_state <- proder %>%
  group_by(ent) %>%
  summarise(skin_state_pop = mean(proder_label, na.rm = TRUE))


#mean(proder$proder_label, na.rm = TRUE) # Mean skin tone is 7.08

```


# Join datasets

```{r join}
# Join all datasets
cand_skin_results <- bind_rows(dep_cand_results, sen_cand_results, pres_cand_results)

# Bring contest column after name column
cand_skin_results %<>%
  select(id, name, contest, everything())

```


```{r new_vars}

# Create skin_pp, which imputes skin tone data from ine to sv if sv is absent.
cand_skin_results %<>%
  mutate(skin_pp = ifelse(is.na(skin_sv), skin_ine, skin_sv))

# Create new skin tone variables
# Add mean skin tone per state, match by ent and state_id
cand_skin_results %<>%
  left_join(proder_state, by = c("state_id" = "ent"))


# Mean candidate skin tone per state
cand_skin_results %<>%
  group_by(state_id) %>%
  mutate(skin_state_can = mean(skin_pp, na.rm = TRUE)) %>%
  ungroup()

# Mean candidate skin tone per contest
cand_skin_results %<>%
  group_by(contest) %>%
  mutate(skin_contest_can = mean(skin_pp, na.rm = TRUE)) %>%
  ungroup()

# Mean candidate skin tone per contest and state
cand_skin_results %<>%
  group_by(state_id, contest) %>%
  mutate(skin_contest_state_can = mean(skin_pp, na.rm = TRUE)) %>%
  ungroup()

# Mean candidate skin tone per party
cand_skin_results %<>%
  group_by(party) %>%
  mutate(skin_party_can = mean(skin_pp, na.rm = TRUE)) %>%
  ungroup()


# Skin tone difference
cand_skin_results$skin_diff <- (cand_skin_results$skin_pp - cand_skin_results$skin_state_pop)^2 %>% sqrt()

```


# Export

```{r export}
# Export the dataset
write_csv(cand_skin_results, "../data/processed/05_cand_skin_results.csv")

write_csv(palette, "../data/processed/05_palette.csv")

write_csv(proder, "../data/processed/05_proder.csv")

```
