==> code/06_description.Rmd <==
---
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../products") })

title: "Descriptive analysis"
author: "Silvia Téliz"
date: "2025-05-13"
output:
  html_document:
    toc: true
    toc_depth: 2      
    toc_float: true
---

This script produces the descriptive analysis of the main dataset (`data/processed/cand_skin_results.Rmd`).


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE,  include = TRUE, warning = FALSE, message = FALSE)
```


```{r packages}
library(tidyverse)
library(magrittr)
library(readr)
library(ggplot2)
library(ggrepel)
library(scales)
library(sf)
library(grid)
library(gridExtra)
library(kableExtra)
library(gt)
library(gtExtras)
```


```{r data}
data <- read_csv("../data/processed/05_cand_skin_results.csv")
proder <- read_csv("../data/processed/05_proder.csv")
palette <- read_csv("../data/processed/05_palette.csv") %>%
  arrange(label)
```



# Image data availability

```{r t_missing_all}
# Comparison of missing data by source
missing_all <- data %>%
  group_by(contest, ine_d, sv_d) %>%
  summarize(n = n()) %>%
  mutate(miss_class = case_when(
    ine_d == 0 & sv_d == 0 ~ "Missing in both sources",
    ine_d == 1 & sv_d == 1 ~ "Existing in both sources",
    ine_d == 0 ~ "Missing in INE",
    sv_d == 0 ~ "Missing in SaberVotar")) %>%
  group_by(contest) %>%
  mutate(total_contest = sum(n),
         total = 1242)

# Reorder contest factor levels
missing_all$contest <- factor(missing_all$contest, 
                               levels = c("pres", "sen", "dep"),
                               labels = c("Presidency", "Senate", "Congress"))

# Add proportion
missing_all <- missing_all %>%
  mutate(prop = round(n / total_contest, 2)) %>%
  ungroup() %>%
  mutate(miss_class = factor(miss_class, 
                             levels = c("Existing in both sources", 
                                        "Missing in both sources",
                                        "Missing in INE", 
                                        "Missing in SaberVotar")))

# Total number of candidates with missing data in both sources
n_missing <- missing_all %>%
  group_by(miss_class) %>%
  summarize(n = sum(n),
            prop = round((n / 1242)*100, 2))
 
```

**Main takeaways:**

- A total of 21 candidates (1.69%) are missing data from both, their INE profile and SaberVotar profile. These candidates are from the MC party and the PAN-PRI-PRD coalition. They are part of the congress race and senate race. They belong to the states: Mexico, Jalisco, Guanajuato, San Luis Potosi, Coahuila, Oaxaca, Mexico City, Chiapas, Quintana Roo, and Guerrero.
- A total of 144 candidates (11.59%) are missing data from their INE profile.
- A total of 32 candidates (2.58%) are missing data from their SaberVotar profile.
- The SaberVotar data has a lower proportion of missing data. Therefore, a possible solution is to use this data source and impute skin tone measurements from the INE profile picture if needed. Averaging both measures would not be optimal because skin tone difference from both sources is often large.

**Correlations with outcome variable:**
- The existence of an INE profile picture is weakly and positively correlated with electoral success (r=`r round(cor(data$ine_d, data$win, use = "pairwise.complete.obs"), 2)`). 
- The existence of a SaberVotar profile picture is weakly and positively correlated with electoral success (r=`r round(cor(data$sv_d, data$win, use = "pairwise.complete.obs"), 2)`). 
- The existence of a social media profile (a rough proxy for presence in public media environment) is weakly and positively correlated with electoral success (r=`r round(cor(data$social_media_d, data$win, use = "pairwise.complete.obs"), 2)`). 

```{r p_missing_all}
# Plot missing data by source
ggplot(missing_all, aes(x = contest, y = n, fill = miss_class)) +
  geom_bar(stat = "identity", position = "fill") +
  labs(title = "Image data availability, by source and contest",
       x = "Contest",
       y = "Proportion of candidates") +
  theme_classic() +
  theme(plot.margin = margin(10, 5, 10, 10),
        panel.grid.major.x = element_line(color = "white", size = 0.7),
        plot.title = element_text(hjust = 0, 
                                  face = "bold", 
                                  margin = margin(b = 25))) +
  # Change legend title
  labs(fill = "Profile picture availability") +
  scale_fill_manual(values = c("Existing in both sources" = "gray90",
                               "Missing in both sources" = "black",
                               "Missing in INE" = "gray50",
                               "Missing in SaberVotar" = "gray70")) +
  scale_y_continuous(labels = percent)
  #coord_flip()
  

```




## Distribution of missing data {.tabset}
This section shows the distribution candidates with missing visual data from both sources, INE and SaberVotar. 

- A total of 21 candidates (1.69%) are missing profile pictures from both, their INE profile and SaberVotar profile. 
- Out of all candidates with missing data from both sources, 18 are female (85.71%) and 3 male (14.28%).  
- Candidates with missing data from both sources are part of the congress race and senate race. 
- These candidates belong to the states: Mexico, Jalisco, Guanajuato, San Luis Potosi, Coahuila, Oaxaca, Mexico City, Chiapas, Quintana Roo, and Guerrero.
- These candidates are from the MC party and the PAN-PRI-PRD coalition. 


### By state

```{r t_missing_state, echo = F}
# Missing image data for all contests, by state
missing_state <- data %>%
  group_by(state) %>%
  mutate(na_both = ifelse(ine_d == 0 & sv_d == 0, 1, 0)) %>%
  summarize(n_ine = sum(ine_d == 0),
            prop_ine = round(n_ine / n(), 2),
            n_sv = sum(sv_d == 0),
            prop_sv = round(n_sv / n(), 2),
            n_both = sum(na_both ==1),
            prop_both = round(n_both / n(), 2)) %>%
  arrange(desc(prop_both))

# Capitalize first letter of "state"
missing_state$state <- str_to_title(missing_state$state)

# Recode Ciudad de Mexico
missing_state$state %<>% as_factor() %>%
  fct_recode("Mexico City" = "Ciudad De Mexico")

```


```{r t_mex_map, include = F}
mex_map <- st_read("mex.geojson")%>%
  # change first column name
  mutate(state = shapeName)

# Change state names
mex_map$state %<>% as_factor() %>%
  fct_recode("Mexico City" = "Distrito Federal",
             "Coahuila" = "Coahuila de Zaragoza",
                      "Michoacan" = "Michoacan de Ocampo",
                      "Queretaro" = "Queretaro de Arteaga",
                      "Veracruz" = "Veracruz de Ignacio de la Llave")

map_data <- mex_map %>%
  left_join(missing_state, by = "state") 

```


```{r p_map}
# Map labels
map_labels <- map_data %>%
  filter(prop_both > 0) %>%
  mutate(
    centroid = st_centroid(geometry),
    lon = st_coordinates(centroid)[, 1],
    lat = st_coordinates(centroid)[, 2],
    label = state)

# Map
ggplot(map_data) +
  geom_sf(aes(fill = prop_both), color = "white", size = 0.2) +
  scale_fill_gradient(
    low = "gray90",
    high = "black",
    na.value = "red",
    name = "Proportion of candidates\nwith missing image data",
    labels = label_percent(accuracy = 1)) +
  labs(title = "Distribution of candidates with missing image data from both sources") +
  geom_text_repel(
    data = map_labels,
    aes(x = lon, y = lat, label = label),
    size = 2,
    nudge_y = 0,
    nudge_x = -8,
    direction = "both",
    max_iter = 2000,
    segment.color = "black",
    segment.size = 0.2,
    min.segment.length = 0.2,
    max.overlaps = Inf) +
  theme_void() +  # removes axes, ticks, and background
  theme(legend.position = "right",          
    plot.title = element_text(hjust = 0, face = "bold")) +
  theme(plot.margin = margin(10, 25, 10, 20))
  


```


```{r t_state}
# Missing data per state and contest
missing_state_contest <- data %>%
  group_by(state, contest) %>%
  mutate(na_both = ifelse(ine_d == 0 & sv_d == 0, 1, 0)) %>%
  summarize(n_both = sum(na_both ==1),
            prop_both = round((n_both / n())*100, 2),
            n_total = n()) %>%
  pivot_wider(names_from = contest, values_from = c(n_both, prop_both, n_total), 
              names_sep = "_")  %>% 
  select(-c(n_both_pres, prop_both_pres, n_total_pres)) %>%
  mutate(total_n = n_both_sen + n_both_dep,
         total_prop = round((total_n / (n_total_sen + n_total_dep))*100, 2)) %>%
  arrange(desc(total_prop)) %>%
  # Remove last row
  filter(!is.na(total_prop))

# Rearrange columns
missing_state_contest %<>%
  select(state, n_both_sen, prop_both_sen, n_both_dep, prop_both_dep, total_n, total_prop)

# Capitalize first letter of "state"
missing_state_contest$state %<>% str_to_title()

# Recode Ciudad de Mexico
missing_state_contest$state %<>% as_factor() %>%
  fct_recode("Mexico City" = "Ciudad De Mexico")

# Print missing_state with kable
missing_state_contest %>%
  rename(
    State = 1,
    "Senate (n)" = 2,
    "Senate (%)" = 3,
   "Congress (n)" = 4,
    "Congress (%)" = 5,
    "All contests (n)" = 6,
    "All contests (%)" = 7) %>%
  kable() %>%
  kable_styling("striped", full_width = T) %>%
  add_header_above(c("Distribution of candidates with missing image data, by state and contest" = 7)) %>% 
  scroll_box(height = "300px")
```

### By gender

```{r t_missing_gender}
# Missing image data for all contests, by state
missing_gender <- data %>%
  group_by(contest, sex, ine_d, sv_d) %>%
  summarize(n = n()) %>%
  mutate(miss_class = case_when(
    ine_d == 0 & sv_d == 0 ~ "Missing in both sources",
    ine_d == 1 & sv_d == 1 ~ "Existing in at least one source",
    ine_d == 0 ~ "Existing in at least one source",
    sv_d == 0 ~ "Existing in at least one source")) %>%
  group_by(contest) %>%
  mutate(total_contest = sum(n),
         total = 1242)

# Reorder contest factor levels
missing_gender$contest <- factor(missing_gender$contest, 
                               levels = c("pres", "sen", "dep"),
                               labels = c("Presidency", "Senate", "Congress"))

# Add proportion
missing_gender %<>%
  mutate(prop = round(n / total_contest, 2)) %>%
  ungroup() %>%
  mutate(miss_class = factor(miss_class, 
                             levels = c("Existing in at least one source", 
                                        "Missing in both sources")))

# Recode sex
missing_gender$sex %<>% as_factor() %>%
  fct_recode("Female" = "f",
             "Male" = "m",
             "Non-binary" = "nb")

```

```{r p_missing_gender}
# Plot missing data by gender and contest
ggplot(missing_gender, aes(x = contest, y = n, fill = miss_class)) +
  geom_bar(stat = "identity", position = "fill") +
  geom_text(data = ~subset(., miss_class == "Missing in both sources"),
    aes(label = paste0("n = ", n)),
    position = position_fill(vjust = 0.1),
    color = "black",
    size = 3) +
  facet_wrap(vars(sex)) +
  labs(title = "Image data availability, by gender and contest",
       x = "Contest",
       y = "Proportion of candidates") +
  theme_classic() +
  theme(plot.margin = margin(10, 5, 10, 10),
        panel.grid.major.x = element_line(color = "white", size = 0.7),
        plot.title = element_text(hjust = 0, 
                                  face = "bold", 
                                  margin = margin(b = 25)),
        axis.text.x = element_text(angle = 45, hjust=1)) +
  # Change legend title
  labs(fill = "Profile picture availability") +
  scale_fill_manual(values = c("Existing in at least one source" = "gray90",
                               "Missing in both sources" = "gray30")) +
  scale_y_continuous(labels = percent)
```



### By party

```{r t_missing_party}
# Missing image data for all contests, by state
missing_party <- data %>%
  group_by(contest, party, ine_d, sv_d) %>%
  summarize(n = n()) %>%
  mutate(miss_class = case_when(
    ine_d == 0 & sv_d == 0 ~ "Missing in both sources",
    ine_d == 1 & sv_d == 1 ~ "Existing in at least one source",
    ine_d == 0 ~ "Existing in at least one source",
    sv_d == 0 ~ "Existing in at least one source")) %>%
  group_by(contest) %>%
  mutate(total_contest = sum(n),
         total = 1242)

# Reorder contest factor levels
missing_party$contest <- factor(missing_party$contest, 
                               levels = c("pres", "sen", "dep"),
                               labels = c("Presidency", "Senate", "Congress"))

# Add proportion
missing_party %<>%
  mutate(prop = round(n / total_contest, 2)) %>%
  ungroup() %>%
  mutate(miss_class = factor(miss_class, 
                             levels = c("Existing in at least one source", 
                                        "Missing in both sources")))

# Recode sex
missing_party$party %<>% as_factor() %>%
  fct_recode("MC" = "mc",
             "PAN-PRI-PRD" = "pan_pri_prd",
             "PT" = "pt",
             "PVEM" = "pvem",
             "PRI" = "pri",
             "PRD" = "prd",
             "PAN" = "pan",
             "MORENA" = "morena",
             "PVEM-PT-MORENA" = "pvem_pt_morena")
```



```{r p_missing_party, echo = F}
# Plot missing data by gender and contest
missing_party %>%
  filter(party == "MC" | party == "PAN-PRI-PRD") %>% # Only keep parties with missing data
  ggplot(aes(x = contest, y = n, fill = miss_class)) +
  geom_bar(stat = "identity", position = "fill") +
  geom_text(data = ~subset(., miss_class == "Missing in both sources"),
    aes(label = paste0("n = ", n)),
    position = position_fill(vjust = 0.1),
    color = "black",
    size = 3) +
    facet_wrap(vars(party)) +
  labs(title = "Image data availability, by party and contest",
       x = "Contest",
       y = "Proportion of candidates",
       caption = "Note: MC and the coalition PAN-PRI-PRD are the only parties in which candidates had mising image data from both sources\n(INE and SaberVotar). The candidates from all other the parties and coalitions (MORENA, PT, PVEM, PVEM-PT-MORENA,\nPRI, PAN, PRD) had profile pictures in at least one source.") +
  theme_classic() +
  theme(plot.margin = margin(10, 5, 10, 10),
        panel.grid.major.x = element_line(color = "white", size = 0.7),
        plot.title = element_text(hjust = 0, 
                                  face = "bold", 
                                  margin = margin(b = 25)),
        axis.text.x = element_text(angle = 45, hjust=1),
        plot.caption = element_text(hjust = 0, size = 9, margin = margin(t = 20)),
        plot.caption.position = "plot") +
  # Change legend title
  labs(fill = "Profile picture availability") +
  scale_fill_manual(values = c("Existing in at least one source" = "gray90",
                               "Missing in both sources" = "gray30")) +
  scale_y_continuous(labels = percent)

```


# Skin tone distribution 

```{r skin_missing}
missing_skin <- data %>%
  filter(ine_d == 1 | sv_d == 1) %>%
  summarize(n_missing = sum(is.na(skin_pp)),
            prop_existing = round(n_missing / n(), 5),
            prop_total = round(n_missing / nrow(data), 5))

```


**Main takeaways:**

- Skin tone measure is the PRODER scale, which ranges from 1 (lightest tone) to 11 (darkest tone).
- The skin tone classification algorithm was not able to measure the skin tone of `r missing_skin$n_missing` candidates. This number represents `r round(missing_skin$prop_existing*100, 2)`% of the candidates with image data and `r round(missing_skin$prop_total*100, 2)`% of the total candidates in the dataset. This issue may be resolved by adjusting the algorithm's facial detection parameters. For now, the data imputation strategy is to use the SaberVotar data and impute measures from the INE data whenever necessary.
- The candidate skin tone distribution is slightly right-skewed, indicating that most candidates have a medium-light skin tone. The mean skin tone is `r round(mean(data$skin_pp, na.rm = TRUE), 2)` and the standard deviation is `r round(sd(data$skin_pp, na.rm = TRUE), 2)`. The skin tone distribution is shown in the first figure.
- The national population mean skin tone is `r round(mean(proder$proder_label),2)` with a standard deviation of `r round(sd(proder$proder_label),2)`, according to PRODER data. Therefore, the mean skin tone of candidates is below the national average by 2.18 points.

## Candidates' skin tone {.tabset}

### Overall distribution

```{r skin_hist_cand}
data %>%
  filter(!is.na(skin_pp)) %>%
  ggplot(aes(x = factor(skin_pp))) +  
  geom_bar(aes(fill = factor(skin_pp)), color = "white") +
  scale_fill_manual(
    values = palette$hex,
    name = "PRODER scale") +
  geom_vline(xintercept = 4.86, linetype = "solid", color = "black", size = 0.5) +
  annotate("text", x = 4.8, y = 270, label = "Mean = 4.86", vjust = -2, hjust = -0.3, size = 3) +
  annotate("text", x = 4.8, y = 265, label = "SD = 1.92", vjust = -.5, hjust = -0.3, size = 3) +
  labs(
    title = "Candidate skin tone distribution",
    x = "Skin tone",
    y = "Count") +
  scale_y_continuous(expand = c(0, 0), limit = c(0, 300)) +
  theme_classic() +
  theme(plot.margin = margin(10, 5, 10, 10),
        panel.grid.major.x = element_line(color = "white", size = 0.7),
        plot.title = element_text(hjust = 0, 
                                  face = "bold", 
                                  margin = margin(b = 30)))
```


### By state

```{r skin_state, include=FALSE}
# Grouped summary + raw data column
skin_state <- data %>%
  filter(state != "nacional") %>%
  group_by(state) %>%
  summarize(
    n = n(),
    mean_skin_cand = round(mean(skin_pp, na.rm = TRUE), 2),
    sd_skin_cand = round(sd(skin_pp, na.rm = TRUE), 2),
    mean_skin_pop = round(mean(skin_state_pop), 2),
    sd_skin_pop = round(sd(skin_state_pop), 2),
    skintone_dist = list(skin_pp)) 

# Capitalize first letter of "state"
skin_state$state %<>% str_to_title()

# Recode Ciudad de Mexico
skin_state$state %<>% as_factor() %>%
  fct_recode("Mexico City" = "Ciudad De Mexico")

# Join with map data
map_data <- mex_map %>%
  left_join(skin_state, by = "state")

# Change column names
skin_state <- skin_state %>%
  rename(State = state,
         Mean = mean_skin_cand,
         SD = sd_skin_cand,
         Distribution = skintone_dist)

```


```{r map_skin_cand}
ggplot(map_data) +
  geom_sf(aes(fill = mean_skin_cand), color = "white", size = 0.1) +
  scale_fill_gradientn(
    colors = c("#b69279", "#b1886c", "#a87f64"),
    limits = c(4, 5.5),
    name = "PRODER scale",
    guide = guide_colorbar(
      ticks = TRUE,
      ticks.colour = "black",
      frame.colour = "white")) +
  labs(title = "Candidate skin tone distribution",
       caption = "Note: The map is colored according to the mean skin tone of candidates in each state.") +
  theme_void() + 
  theme(legend.position = "right",          
    panel.grid = element_blank(),
    plot.title = element_text(hjust = 0, face = "bold"),
    plot.caption = element_text(hjust = 0, size = 9, margin = margin(t = 15))) +
   annotation_custom(
    grob = linesGrob(y = unit(c(0.01, 0.01), "npc"), gp = gpar(col = "black", lwd = 0.5)),
    ymin = -Inf, ymax = Inf, xmin = -Inf, xmax = Inf) +
  theme(plot.margin = margin(10, 25, 10, 20))
 

```

```{r t_skin_state}
# Table
skin_state %>%
  select(c(State, Mean, SD, Distribution)) %>%
  gt() %>%
  tab_header(title = "Candidate skin tone distribution by state") %>%
  gtExtras::gt_plt_dist(
    column = Distribution,
    type = "boxplot",
    fill = "gray70",
    bw = 0.5)
```




### By contest
```{r skin_contest}
# Grouped summary + raw data column
skin_contest <- data %>%
  group_by(contest) %>%
  summarize(
    n = n(),
    mean = round(mean(skin_pp, na.rm = TRUE), 2),
    sd = round(sd(skin_pp, na.rm = TRUE), 2),
    skintone_dist = list(skin_pp)) %>%
  arrange(desc(mean))

# Recode contest
skin_contest$contest %<>% as_factor() %>%
  fct_recode("Congress" = "dep",
             "Senate" = "sen",
             "Presidency" = "pres")

skin_contest %>%
  select(-c(n)) %>%
  rename(
    Contest = contest,
    Mean = mean,
    SD = sd,
    Distribution = skintone_dist) %>%
  gt() %>%
  tab_header(title = "Candidate skin tone distribution by contest") %>%
  gtExtras::gt_plt_dist(
    column = Distribution,
    type = "boxplot",
    fill = "gray70",
    bw = 0.5)


```



### By party
```{r skin_party}
# Grouped summary + raw data column
skin_party <- data %>%
  filter(party != "ind") %>%
  group_by(party) %>%
  summarize(
    n = n(),
    mean = round(mean(skin_pp, na.rm = TRUE), 2),
    sd = round(sd(skin_pp, na.rm = TRUE), 2),
    skintone_dist = list(skin_pp)) %>%
  arrange(desc(mean))

# Recode party
skin_party$party %<>% as_factor() %>%
  fct_recode("MC" = "mc",
             "PAN-PRI-PRD" = "pan_pri_prd",
             "PT" = "pt",
             "PVEM" = "pvem",
             "PRI" = "pri",
             "PRD" = "prd",
             "PAN" = "pan",
             "MORENA" = "morena",
             "PVEM-PT-MORENA" = "pvem_pt_morena")


skin_party %>%
  select(-c(n)) %>%
  rename(
    Party = party,
    Mean = mean,
    SD = sd,
    Distribution = skintone_dist) %>%
  gt() %>%
  tab_header(title = "Candidate skin tone distribution by party") %>%
  gtExtras::gt_plt_dist(
    column = Distribution,
    type = "boxplot",
    fill = "gray70",
    bw = 0.5)

```




## Population's skin tone {.tabset}

### Overall distribution

```{r skin_hist_pop}
# Plot  candidate skin tone distribution 
proder %>%
  ggplot(aes(x = factor(proder_label))) +
  geom_bar(aes(fill = factor(proder_label)), color = "white") +
  scale_fill_manual(
    values = palette$hex,
    name = "PRODER scale") +
  geom_vline(xintercept = 7.08, linetype = "solid", color = "black", size = 0.5) +
  annotate("text", x = 5.5, 
           y = 1500, label = "Mean = 7.08", 
           vjust = -2, hjust = 0, size = 3) +
  annotate("text", x = 5.5, y = 1500, 
           label = "SD = 2.09", 
           vjust = -.5, hjust = 0, size = 3) +
  labs(title = "Population skin tone distribution",
       x = "Skin tone",
       y = "Count") +
  scale_y_continuous(expand = c(0, 0), limit = c(0, 2000)) +
  theme_classic() +
  theme(plot.margin = margin(10, 5, 10, 10),
        panel.grid.major.x = element_line(color = "white", size = 0.7),
        plot.title = element_text(hjust = 0, 
                                  face = "bold", 
                                  margin = margin(b = 30)))
```

### By state


```{r map_skin_pop}

map_data %>%
  filter(state != "nacional") %>%
ggplot() +
  geom_sf(aes(fill = mean_skin_pop), color = "white", size = 0.1) +
  scale_fill_gradientn(
    colors = palette$hex[4:9],
    limits = c(4, 9),
    name = "PRODER scale",
    guide = guide_colorbar(
      ticks = TRUE,
      ticks.colour = "black",
      frame.colour = "white")) +
  labs(title = "Population skin tone distribution",
       caption = "Note: The map is colored according to the mean skin tone of the population in each state.") +
  theme_void() + 
  theme(legend.position = "right",          
    panel.grid = element_blank(),
    plot.title = element_text(hjust = 0, face = "bold"),
    plot.caption = element_text(hjust = 0, size = 9, margin = margin(t = 15))) +
   annotation_custom(
    grob = linesGrob(y = unit(c(0.01, 0.01), "npc"), gp = gpar(col = "black", lwd = 0.5)),
    ymin = -Inf, ymax = Inf, xmin = -Inf, xmax = Inf) +
  theme(plot.margin = margin(10, 25, 10, 20))
 
```
