==> code/07_analysis.Rmd <==
---
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../output") })
  
title: "Analysis"
author: "Silvia Téliz"
date: "2025-05-13"
output:
  html_document:
    toc: true
    toc_depth: 2      
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r packages}
library(tidyverse)
library(magrittr)
library(modelsummary)
```

```{r data}
data <- read_csv("../data/processed/05_cand_skin_results.csv")

proder <- read_csv("../data/processed/05_proder.csv")
```

# Regression models

## Replication of Campos‐Vazquez & Rivas‐Herrera, 2021

```{r m1}
m1a <- glm(win ~ skin_pp + sex + party + contest,
    data = data, family = "binomial") 

m1b <- lm(prop_votes ~ skin_pp + sex + party + contest,
    data = data)

m1 <- list(m1a, m1b)
```


```{r m2}
m2a <- glm(win ~ skin_pp + skin_diff + sex + party + contest + incumbency,
    data = data, family = "binomial")

m2b <- lm(prop_votes ~ skin_pp + skin_diff + sex + party + contest + incumbency,
    data = data)

m2 <- list(m2a, m2b)

```


```{r export}

# Model 1: replication of Campos‐Vazquez & Rivas‐Herrera, 2021
modelsummary(m1, output = "../output/m1.tex", stars = TRUE, fmt = fmt_decimal(3))


# Model 2: includes incumbency
modelsummary(m2, output = "../output/m2.tex", stars = TRUE, fmt = fmt_decimal(3))


```
